refactor: Created the Utilities directory
This dir will contain extra bits of code that have been
created as a utility to make things more compact in
other parts of the code.
Also helper code that may need to be used in multiple
places can be included in this dir.
