+++ KerbalSimpit/Utilities/Utilities.cs
