+++ Main.csproj
@@ -43,7 +43,7 @@
     <Compile Include="KerbalSimpit\KerbalSimpitPackets.cs" />
     <Compile Include="KerbalSimpit\External\ARPWrapper.cs" />
     <Compile Include="KerbalSimpit\Providers\*.cs" />
-    <Compile Include="KerbalSimpit\Utilities.cs" />
+    <Compile Include="KerbalSimpit\Utilities\*.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="KerbalSimpit\Console\*.cs"/>
     <Compile Include="KerbalSimpit\Localisation_Libs\*.cs"/>
